feat(trusty-mpm): standard harness-agnostic inject_text/observe/summarize on SessionControl (#1461) (#1463)
* feat(trusty-mpm): standard harness-agnostic inject_text/observe/summarize on SessionControl (closes #1461)

Extend the SessionControl trait with three harness-agnostic verbs so ANY
runtime (Claude Code, tcode, future SDK backends) is driven through one
abstraction:

- Submit enum (Enter | NoSubmit | Interrupt) modelling the three keystroke
  intents without leaking tmux argv into the seam.
- RawObservation / Summary value types (Summary reuses the existing
  ActivityState enum — no duplicate state machine).
- inject_text(submit), observe (ALWAYS available, NO LLM), and summarize
  (LLM-OPTIONAL: degrades to Unknown/None when no OPENROUTER_API_KEY, serves
  cached verdicts on content-hash hits via the existing ActivityMonitor).

Wiring (minimal/localized for a trivial rebase against #1458):
- ManagedTmuxDriver gains send_keys_literal + send_interrupt with default
  impls (six existing impls compile untouched); RealTmuxDriver overrides both
  with production-correct tmux argv (literal send-keys / C-c).
- SessionManager::inject dispatches by Submit; SessionManager::observe bundles
  pane + runtime_active + pending fields, LLM-free.
- DaemonSessionControl (real) implements all three; mock_control + the two
  sm_stdio test mocks implement them too.

Back-compat: send/send_input keep their literal+Enter behaviour, asserted
equivalent to inject(.., Enter).


* fix(trusty-mpm): explicit-error defaults for send_keys_literal/send_interrupt + drop unsafe env mutation in test + checked lines cast (refs #1461)

Address trusty-review BLOCK fixes on PR #1463 (ControllableSession API):

- Default ManagedTmuxDriver::send_keys_literal now returns an explicit
  TmuxUnavailable error instead of delegating to send_line (which would
  append Enter and submit a Submit::NoSubmit line — silent-wrong-behavior).
- Default ManagedTmuxDriver::send_interrupt now returns an explicit
  TmuxUnavailable error instead of a silent Ok(()) no-op (Interrupt STOPS a
  task; a no-op would leave it running). Real RealTmuxDriver overrides keep
  the correct literal-no-Enter / C-c behavior; the recording test driver
  overrides both to record, so the NoSubmit/Interrupt dispatch tests still
  prove the right keys.
- observe_returns_raw_pane_without_llm: removed the unsafe
  std::env::remove_var("OPENROUTER_API_KEY") (UB under the multi-threaded
  test runner); observe is LLM-free so the mutation was pointless.
- DaemonSessionControl::observe: checked u32::try_from(lines) instead of
  `lines as u32`, mapping overflow to a Backend error.
- Hoisted Submit into manager.rs module imports (dropped redundant inline
  use); added proper use imports in sm_stdio/tests.rs for the mock impls.


* fix(trusty-mpm): no activity bump on Interrupt + align observe usize signature + approx-float test (review follow-up, refs #1461)

- inject(): only bump last_activity_at for Submit::Enter/NoSubmit; an
  Interrupt (C-c) is a STOP signal, not forward progress, so bumping it
  would mislead the idle/orphan-GC reconciliation. Add a dedicated test
  asserting Interrupt leaves last_activity_at None while Enter/NoSubmit
  set it.
- Align the observe line-count to usize end-to-end: ManagedTmuxDriver::
  capture, SessionManager::observe, and capture_pane now take usize,
  removing the load-bearing u32::try_from narrowing in
  DaemonSessionControl::observe. The single u32 conversion the tmux -S
  argv requires is confined to the RealTmuxDriver edge.
- verdict_to_summary_keeps_real_summary: use an approx float comparison
  (abs diff < 1e-6) instead of assert_eq! on an f64.


---------

+++ crates/trusty-mpm/src/supervisor/tests.rs
@@ -78,7 +78,7 @@ impl ManagedTmuxDriver for FakeTmux {
         Ok(())
     }
 
-    fn capture(&self, _name: &str, _lines: u32) -> Result<String, ManagedError> {
+    fn capture(&self, _name: &str, _lines: usize) -> Result<String, ManagedError> {
         Ok("working on the task...\n$ cargo test".to_owned())
     }
