+++ crates/trusty-mpm/tests/orphan_gc_sweep.rs
@@ -44,7 +44,7 @@ impl ManagedTmuxDriver for RecordingDriver {
     fn send_line(&self, _name: &str, _text: &str) -> Result<(), ManagedError> {
         Ok(())
     }
-    fn capture(&self, _name: &str, _lines: u32) -> Result<String, ManagedError> {
+    fn capture(&self, _name: &str, _lines: usize) -> Result<String, ManagedError> {
         Ok(String::new())
     }
     fn list_sessions(&self) -> Result<Vec<String>, ManagedError> {
